refactor: Rename navigation routes for clarity
This commit refactors the navigation route data objects for better clarity and consistency. The `@SerialName` annotations have been removed, and routes like `MoviesRoute` and `TvShowsRoute` are now renamed to `MoviesListRoute` and `TvShowsListRoute` respectively.

This change is propagated across the application, updating route imports and usages in navigation graphs, top-level destinations, and the main `AppState`. The goal is to make the purpose of each route more explicit in the code.